fixing compiler warning, "comparison is always true due to limited range of data type"
git-svn-id: svn://fly.thruhere.net/forsaken@2013 1361c5d3-c82f-4d94-b015-8aedb5f389d9

+++ ProjectX/trunk/src/aifollow.c
@@ -170,7 +170,7 @@ void AI_AIR_FOLLOWPATH( register ENEMY * Enemy )
 								&Enemy->Object.Pos, &TempVector, &TempForwardVector, &TempUpVector,
 								&TempVector, EnemyTypes[Enemy->Type].SecondaryWeaponType, FALSE );
 
-					if( MineIndex != -1 )
+					if( MineIndex != (uint16) -1 )
 					{
 						SecBulls[MineIndex].LifeSpan = 10.0F * 60.0F;
 					}
